API/requirements.txt
asgiref==3.7.2
Django==4.2.5
django-cors-headers==4.3.1
django-crontab==0.7.1
django-model-utils==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
gunicorn==21.2.0
packaging==23.2
pillow==10.2.0
psycopg2-binary==2.9.7
PyJWT==2.8.0
python-decouple==3.8
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
whitenoise==6.6.0
